Move the import related functionality to one place
