+++ compiler/hie-core/src/Development/IDE/Import/DependencyInformation.hs
@@ -1,7 +1,7 @@
 -- Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 -- SPDX-License-Identifier: Apache-2.0
 
-module Development.IDE.Functions.DependencyInformation
+module Development.IDE.Import.DependencyInformation
   ( DependencyInformation(..)
   , RawDependencyInformation(..)
   , NodeError(..)
